pyproject.toml
[project]
name = "esctl"
version = "1.13.0"
description = "CLI tool to run commands on ES clusters"
authors = [{ name = "Simon Bordeyne", email = "simon.bordeyne@happn.fr" }]
requires-python = ">=3.11,<4.0"
readme = "README.md"
license = "Apache-2.0"
keywords = [
    "elasticsearch",
    "cli",
    "kubernetes",
    "gce",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "rich >=14.2.0",
    "requests >=2.32.5,<3",
    "typer >=0.19.2",
    "jmespath >=1.0.1,<2",
    "pydantic >=2.12.3",
    "jsonpath-ng >=1.7.0,<2",
    "kubernetes >=34.1.0",
    "ipython >=9.6.0",
    "blake3 >=1.0.8",
    "orjson >=3.11.3,<4",
    "elasticsearch8",
    "elasticsearch9",
    "kubernetes >=34.1.0,<35",
    "ruamel.yaml >=0.18.15",
]
license-files = ["LICENSE"]

[project.scripts]
esctl = "esctl.cli:app"

[dependency-groups]
dev = ["ruff>=0.14.0,<0.15"]
docs = [
    "mkdocs>=1.6.1,<2",
    "pymdown-extensions>=10.16.1,<11",
    "mkdocs-techdocs-core>=1.6.0,<2",
    "mike>=2.1.3,<3",
    "mkdocs-typer2>=0.1.6,<0.2",
    "mkdocs-awesome-nav>=3.2.0,<4",
]

[tool.uv]
package = true
default-groups = [
    "dev",
    "docs",
]
required-version = ">=0.9.2,<1.0.0"
add-bounds = "major"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.mypy]
disable_error_code = ["import-untyped"]
